+++ package.json
@@ -47,7 +47,9 @@
   "devDependencies": {
     "@changesets/cli": "^2.27.0",
     "@changesets/get-github-info": "^0.6.0",
-    "@sku-private/test-utils": "workspace:*",
+    "@sku-private/eslint-plugin": "workspace:*",
+    "@sku-private/testing-library": "workspace:*",
+    "@sku-private/puppeteer": "workspace:*",
     "@swc/core": "^1.3.84",
     "@tsconfig/node-lts": "^18.12.3",
     "@types/jest": "^30.0.0",
