+++ private/puppeteer/index.ts
@@ -1,0 +1,2 @@
+export { getAppSnapshot } from './appSnapshot.ts';
+export { appSnapshotSerializer } from './appSnapshotSerializer.ts';
