+++ private/test-utils/dirContentsToObject.ts
