+++ private/test-utils/package.json
@@ -7,8 +7,6 @@
     "@jest/globals": "^30.0.3",
     "@prettier/sync": "^0.5.2",
     "@types/css": "^0.0.38",
-    "@types/diff": "^7.0.1",
-    "@types/diffable-html": "^5.0.2",
     "@types/hostile": "^1.3.5",
     "@types/klaw": "^3.0.6",
     "@types/serve-handler": "^6.1.4",
@@ -16,8 +14,6 @@
     "@types/webpack-stats-plugin": "^0.3.5",
     "cli-testing-library": "^3.0.1",
     "css": "^3.0.0",
-    "diff": "^7.0.0",
-    "diffable-html": "^5.0.0",
     "get-port": "^7.1.0",
     "hostile": "^1.3.3",
     "klaw": "^4.1.0",
