+++ private/testing-library/vitest.ts
@@ -1,0 +1,2 @@
+// Re-export vitest setup from cli-testing-library
+import 'cli-testing-library/vitest';
