--- test-utils/vitest/index.ts
@@ -1,1 +1,0 @@
-export { getAppSnapshot } from './appSnapshot.ts';
