--- test-utils/vitest/setup.ts
@@ -1,5 +1,0 @@
-import { expect } from 'vitest';
-import appSnapshotSerializer from './appSnapshotSerializer.ts';
-import 'cli-testing-library/vitest';
-
-expect.addSnapshotSerializer(appSnapshotSerializer);
