+++ tests/multiple-routes.test.ts
@@ -7,7 +7,7 @@ import {
   waitForUrls,
   createCancelSignal,
 } from '@sku-private/test-utils';
-import { getAppSnapshot } from '@sku-private/vitest-utils';
+import { getAppSnapshot } from '@sku-private/puppeteer';
 
 import { createRequire } from 'node:module';
