+++ tests/suspense.test.ts
@@ -1,5 +1,5 @@
 import { describe, beforeAll, afterAll, it } from 'vitest';
-import { getAppSnapshot } from '@sku-private/vitest-utils';
+import { getAppSnapshot } from '@sku-private/puppeteer';
 import path from 'node:path';
 import {
   dirContentsToObject,
