+++ vitest-setup.ts
@@ -1,0 +1,5 @@
+import { expect } from 'vitest';
+import { appSnapshotSerializer } from '@sku-private/puppeteer';
+import '@sku-private/testing-library/vitest';
+
+expect.addSnapshotSerializer(appSnapshotSerializer);
